Tone down README
Dropwizard is not as fast as initially believed. The integration tests for Dropwizard were using ResourceTestRule which does not instantiate the entire application or issue real HTTP requests. Switching to a Jersey based test revealed that Dropwizard is still faster, about twice as fast, but capable of handling 300k requests not 700k out of the box.